+++ metrics-servlets/src/test/java/com/codahale/metrics/servlets/MetricsServletTest.java
@@ -1,9 +1,8 @@
 package com.codahale.metrics.servlets;
 
 import com.codahale.metrics.*;
-import com.codahale.metrics.health.HealthCheckRegistry;
-
-import org.eclipse.jetty.testing.ServletTester;
+import org.eclipse.jetty.http.HttpHeader;
+import org.eclipse.jetty.servlet.ServletTester;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -28,7 +27,8 @@ public class MetricsServletTest extends AbstractServletTest {
     @Override
     protected void setUp(ServletTester tester) {
         tester.setAttribute("com.codahale.metrics.servlets.MetricsServlet.registry", registry);
-        tester.addServlet(MetricsServlet.class, "/metrics");
+        tester.addServlet(MetricsServlet.class, "/metrics")
+                    .setInitParameter("com.codahale.metrics.servlets.MetricsServlet.allowedOrigin", "*");
     }
 
     @Before
@@ -58,6 +58,8 @@ public class MetricsServletTest extends AbstractServletTest {
 
         assertThat(response.getStatus())
                 .isEqualTo(200);
+        assertThat(response.get("Access-Control-Allow-Origin"))
+                .isEqualTo("*");
         assertThat(response.getContent())
                 .isEqualTo("{" +
                                    "\"version\":\"3.0.0\"," +
@@ -74,7 +76,7 @@ public class MetricsServletTest extends AbstractServletTest {
                                        "\"m\":{\"count\":1,\"m15_rate\":0.0,\"m1_rate\":0.0,\"m5_rate\":0.0,\"mean_rate\":3333333.3333333335,\"units\":\"events/second\"}},\"timers\":{\"t\":{\"count\":1,\"max\":1.0,\"mean\":1.0,\"min\":1.0,\"p50\":1.0,\"p75\":1.0,\"p95\":1.0,\"p98\":1.0,\"p99\":1.0,\"p999\":1.0,\"stddev\":0.0,\"m15_rate\":0.0,\"m1_rate\":0.0,\"m5_rate\":0.0,\"mean_rate\":1.0E7,\"duration_units\":\"seconds\",\"rate_units\":\"calls/second\"}" +
                                    "}" +
                                "}");
-        assertThat(response.getContentType())
+        assertThat(response.get(HttpHeader.CONTENT_TYPE))
                 .isEqualTo("application/json");
     }
 
@@ -86,6 +88,8 @@ public class MetricsServletTest extends AbstractServletTest {
 
         assertThat(response.getStatus())
                 .isEqualTo(200);
+        assertThat(response.get("Access-Control-Allow-Origin"))
+                .isEqualTo("*");
         assertThat(response.getContent())
                 .isEqualTo(String.format("{%n" +
                                                  "  \"version\" : \"3.0.0\",%n" +
@@ -146,7 +150,7 @@ public class MetricsServletTest extends AbstractServletTest {
                                                  "    }%n" +
                                                  "  }%n" +
                                                  "}"));
-        assertThat(response.getContentType())
+        assertThat(response.get(HttpHeader.CONTENT_TYPE))
                 .isEqualTo("application/json");
     }
 
@@ -160,7 +164,7 @@ public class MetricsServletTest extends AbstractServletTest {
         final MetricsServlet metricsServlet = new MetricsServlet(metricRegistry);
         metricsServlet.init(servletConfig);
  
-        verify(servletConfig, times(3)).getServletContext();
+        verify(servletConfig, times(1)).getServletContext();
         verify(servletContext, never()).getAttribute(eq(MetricsServlet.METRICS_REGISTRY));
     }
 
@@ -176,7 +180,7 @@ public class MetricsServletTest extends AbstractServletTest {
         final MetricsServlet metricsServlet = new MetricsServlet(null);
         metricsServlet.init(servletConfig);
 
-        verify(servletConfig, times(4)).getServletContext();
+        verify(servletConfig, times(1)).getServletContext();
         verify(servletContext, times(1)).getAttribute(eq(MetricsServlet.METRICS_REGISTRY));
     }