+++ docs/source/about/release-notes.rst
@@ -4,6 +4,19 @@
 Release Notes
 #############
 
+.. _rel-3.0.1:
+
+v3.0.1-SNAPSHOT
+===============
+
+* Fixed NPE in ``MetricRegistry#name``.
+* ``ScheduledReporter`` and ``JmxReporter`` now implement ``Closeable``.
+* Fixed cast exception for async requests in ``metrics-jetty9``.
+* Added support for ``Access-Control-Allow-Origin`` to ``MetricsServlet``.
+* Fixed numerical issue with ``Meter`` EWMA rates.
+* Deprecated ``AdminServletContextListener`` in favor of ``MetricsServlet.ContextListener`` and
+  ``HealthCheckServlet.ContextListener``.
+
 .. _rel-3.0.0:
 
 v3.0.0: June 10 2013